requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile --annotation-style=line --output-file=requirements.txt requirements-in.txt
#
adbutils==0.11.0          # via -r requirements-in.txt, uiautomator2
aiofiles==0.7.0           # via -r requirements-in.txt
alas-webapp==0.3.7        # via -r requirements-in.txt
anyio==1.3.1              # via -r requirements-in.txt
apkutils2==1.0.0          # via adbutils
asgiref            # via uvicorn
async-generator==1.10     # via anyio
cached-property==1.5.2    # via uiautomator2
certifi==2021.5.30        # via requests
cffi==1.15.0              # via gevent
chardet==3.0.4            # via requests
cigam==0.0.3              # via apkutils2
click==8.0.1              # via uvicorn
cnocr==1.2.2              # via -r requirements-in.txt
colorama==0.4.5           # via click, logzero, rich, tqdm, uvicorn
commonmark==0.9.1         # via rich
cycler==0.10.0            # via matplotlib
decorator==5.1.0          # via retry
deprecated==1.2.13        # via uiautomator2
deprecation==2.1.0        # via adbutils
filelock==3.3.0           # via uiautomator2
future==0.18.2            # via zerorpc
gevent==21.12.0           # via zerorpc
gluoncv==0.6.0            # via cnocr
graphviz==0.8.4           # via mxnet
greenlet==1.1.2           # via gevent
h11==0.12.0               # via uvicorn
httptools==0.4.0          # via uvicorn
idna==2.6                 # via requests
imageio==2.9.0            # via -r requirements-in.txt
importlib-metadata==4.8.1  # via click, prettytable
inflection==0.5.1         # via -r requirements-in.txt
jellyfish==0.8.8          # via -r requirements-in.txt
kiwisolver==1.3.2         # via matplotlib
logzero==1.7.0            # via uiautomator2
lxml==4.6.3               # via uiautomator2
lz4==3.1.3                # via -r requirements-in.txt
matplotlib==3.4.3         # via gluoncv
msgpack==1.0.3            # via zerorpc
mxnet==1.6.0              # via -r requirements-in.txt, cnocr
numpy==1.16.6             # via -r requirements-in.txt, cnocr, gluoncv, imageio, matplotlib, mxnet, opencv-python, scipy
opencv-python==4.5.3.56   # via -r requirements-in.txt
packaging==20.9           # via deprecation, uiautomator2
pillow==8.3.2             # via -r requirements-in.txt, cnocr, gluoncv, imageio, matplotlib, uiautomator2
portalocker==2.3.2        # via gluoncv
prettytable==2.2.1        # via -r requirements-in.txt
progress==1.6             # via uiautomator2
py==1.10.0                # via retry
pycparser==2.21           # via cffi
pyelftools==0.27          # via apkutils2
pygments==2.12.0          # via rich
pyparsing==2.4.7          # via matplotlib, packaging
pypresence==4.2.1         # via -r requirements-in.txt
python-dateutil==2.8.2    # via matplotlib
python-dotenv==0.19.0     # via uvicorn
pywebio==1.6.2            # via -r requirements-in.txt
pyyaml==5.4.1             # via -r requirements-in.txt, uvicorn
pyzmq==22.3.0             # via -r requirements-in.txt, zerorpc
requests>=2.20.0          # via adbutils, gluoncv, mxnet, uiautomator2
retry==0.9.2              # via adbutils, uiautomator2
retrying==1.3.3           # via -r requirements-in.txt
rich==11.2.0              # via -r requirements-in.txt
scipy==1.4.1              # via -r requirements-in.txt, gluoncv
six==1.16.0               # via adbutils, cycler, python-dateutil, retrying, uiautomator2
sniffio==1.2.0            # via anyio
starlette==0.14.2         # via -r requirements-in.txt
tornado==6.1              # via pywebio
tqdm==4.62.3              # via -r requirements-in.txt, gluoncv
typing-extensions==4.3.0  # via asgiref, importlib-metadata, rich, uvicorn
ua-parser==0.10.0         # via user-agents
uiautomator2==2.16.17     # via -r requirements-in.txt
urllib3==1.22             # via requests
user-agents==2.2.0        # via pywebio
uvicorn[standard]==0.17.6  # via -r requirements-in.txt
watchgod==0.7             # via uvicorn
wcwidth==0.2.5            # via prettytable
websockets==10.0          # via uvicorn
whichcraft==0.6.1         # via adbutils, uiautomator2
wrapt==1.13.1             # via -r requirements-in.txt, deprecated
xmltodict==0.12.0         # via apkutils2
zerorpc==0.6.3            # via -r requirements-in.txt
zipp==3.6.0               # via importlib-metadata
zope.event==4.5.0         # via gevent
zope.interface==5.4.0     # via gevent
psycopg2-binary
cryptography
fastapi
python-multipart

# The following packages are considered to be unsafe in a requirements file:
# setuptools
